producer-app/requirements.txt
kafka-python==2.0.2
watchdog==3.0.0
